refactor: make keydown separate files per key
