+++ src/event/keydown/keydown.ts
@@ -1,0 +1,12 @@
+import { processEnter } from "./enter.ts";
+
+/**
+ * Event listener function for handling key down events in the terminal.
+ *
+ * @param event event listener {@link KeyboardEvent}
+ */
+export function keydown(event: KeyboardEvent) {
+  if (event.key === "Enter") {
+    processEnter(event);
+  }
+}
